Feature/rename archives archival (#129)
* renaming archives to archival. breaking out helpers into separate files

* removing unused file

* more replacing archives

* deleting unused file
