+++ client/routes/domain/workflow-archival/helpers/get-domain.js
@@ -1,0 +1,3 @@
+import get from 'lodash-es/get';
+
+export default domainSettings => get(domainSettings, 'domainInfo.name', '');
