+++ client/routes/domain/workflow-archival/helpers/get-visibility-archival-status.js
@@ -1,0 +1,4 @@
+import get from 'lodash-es/get';
+
+export default domainSettings =>
+  get(domainSettings, 'configuration.visibilityArchivalStatus', '');
