+++ client/routes/domain/workflow-archival/helpers/is-archival-enabled.js
@@ -1,0 +1,6 @@
+import isHistoryArchivalEnabled from './is-history-archival-enabled';
+import isVisibilityArchivalEnabled from './is-visibility-archival-enabled';
+
+export default domainSettings =>
+  isHistoryArchivalEnabled(domainSettings) &&
+  isVisibilityArchivalEnabled(domainSettings);
