+++ client/routes/domain/workflow-archival/helpers/is-history-archival-enabled.js
@@ -1,0 +1,4 @@
+import getHistoryArchivalStatus from './get-history-archival-status';
+
+export default domainSettings =>
+  getHistoryArchivalStatus(domainSettings) === 'ENABLED';
