+++ client/routes/domain/workflow-archival/helpers/is-visibility-archival-enabled.js
@@ -1,0 +1,4 @@
+import getVisibilityArchivalStatus from './get-visibility-archival-status';
+
+export default domainSettings =>
+  getVisibilityArchivalStatus(domainSettings) === 'ENABLED';
