+++ client/routes/domain/workflow-archival/helpers/replace-domain.js
@@ -1,0 +1,4 @@
+import getDomain from './get-domain';
+
+export default (message, domainSettings) =>
+  message.replace(/\{domain\}/, getDomain(domainSettings));
